cephadm: move a pair of systemd unit status funcs to systemd.py
